project/app/serializers.py
from rest_framework import serializers

from project.app.models import ExampleModel


class ExampleSerializer(serializers.ModelSerializer):
    class Meta:
        model = ExampleModel
        fields = '__all__'
