requirements.txt
django==3.2.12
djangorestframework==3.13.1
pyyaml==6.0
uritemplate==4.1.1
